Moved calculate_distances into geometry subfolder of transport folder. Renamed dependencies.
